moved createTaskDictionary
